requirements.txt
toml
pkg_resources
cyclonedx-bom
safety
pytest